+++ backend/aci/common/db/crud/__init__.py
@@ -1,4 +1,5 @@
 from . import (
+    agents,
     app_configurations,
     apps,
     functions,
@@ -11,6 +12,7 @@ from . import (
 )
 
 __all__ = [
+    "agents",
     "app_configurations",
     "apps",
     "functions",
